fix(404-app): return HTTP 404 status by rendering UI from not-found.tsx
The Next.js 404 handler was returning HTTP 200 because app/page.tsx is a
normal route. Move the directory UI into app/not-found.tsx and reduce
page.tsx to a server component that calls notFound(), so every request
reaches Next.js's 404 path and responds with HTTP 404.

This matches the README contract and makes tests/404-handler.test.ts
pass (it asserts response.status === 404).
